Multiple bug fixes
