+++ NearBeach/templates/NearBeach/organisations/organisation_information.html
@@ -8,5 +8,6 @@
     <organisation-modules v-bind:destination="'organisation'"
                           v-bind:location-id="{{organisation_id}}"
                           v-bind:customer-results="{{customer_results}}"
+                          v-bind:title-list="{{title_list}}"
     ></organisation-modules>
 {% endblock %}
